DBUnit을 이용한 DB초기화 일단 끝남 - null 값처리까지함.

+++ src/test/java/com/example/repository/BoardArticleRepositoryTest.java
@@ -68,7 +68,7 @@ public class BoardArticleRepositoryTest {
         BoardArticle getArticle = getJustoneArticle();
         repository.updateNumRead(getArticle.getId());
         getArticle = repository.findOne(getArticle.getId());
-        assertEquals(1, getArticle.getNum_read());
+        assertEquals(1, getArticle.getNum_read().intValue());
 	}
