+++ src/test/java/com/example/repository/BoardCreateTestCode.java
@@ -66,7 +66,7 @@ public class BoardCreateTestCode {
         BoardArticle getArticle = getJustoneArticle();
         repository.updateNumRead(getArticle.getId());
         getArticle = repository.findOne(getArticle.getId());
-        assertEquals(1, getArticle.getNum_read());
+        assertEquals(1, getArticle.getNum_read().intValue());
 	}
